be more robust about ensuring the dynamic .rst files were generated before setup.py AND before sphinx
